+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All Notable changes to `digipolisgent/flanders-basicregisters` package.
 
+## [0.2.2]
+
+### Added
+
+* Added method to PostInfoNames to get all names at once.
+  The post info names object contains all PostInfoName objects at once
+  including the main name and sub-municipality names. There are use cases
+  where an array of all names, starting with the main name, are required.
+
+### Changed
+
+* The main name is written in all caps. That name is now transformed into
+  Capitalized Words.
+
 ## [0.2.1]
 
 ### Changed
@@ -80,6 +94,7 @@ Covered:
 * Added service method to get the list of post info values.
 * Added service method to get the details of a single post info value.
 
+[0.2.2]: https://github.com/digipolisgent/php_package_dg-flanders-basicregisters/compare/0.2.1...0.2.2
 [0.2.1]: https://github.com/digipolisgent/php_package_dg-flanders-basicregisters/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/digipolisgent/php_package_dg-flanders-basicregisters/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/digipolisgent/php_package_dg-flanders-basicregisters/compare/0.1.0...0.1.1